+++ tests/aws_emr_launch/constructs/emr_constructs/test_cluster_configuration.py
@@ -101,7 +101,8 @@ default_config = {
             ]
         },
         'StepConcurrencyLevel': 1
-    }
+    },
+    'OverrideInterfaces': {}
 }